Cleanup, Added Baskin Robbins
I intentionally put the BR flavors in here in a different way than Graeter's (well, it wasn't intentional when I started), but this way we can stare at both options for a hot minute.  BR includes "(Baskin Robbins)" in parentheses AFTER the flavor name, so it sorts ordered by flavor.  Only flavors that weren't already in the list are added.  I'm glad to redo this the other way, or to change Graeter's to match this way in my dubious free time.

Also, I cleaned up a couple of typos and deleted duplicates that creeped in due to case insensitivity.